Relocate numba specific tests in test_utils (and skip if not installed)
